+++ .deploy/restart.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+echo "[deploy] Restarting..."
+pm2 restart nodepress
+echo "[deploy] Restart done."